packages/backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.3
sqlalchemy==2.0.41
alembic==1.15.2
# psycopg[binary] handles async natively, no need for asyncpg
psycopg[binary]==3.3.3
pydantic-settings==2.9.1
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
httpx==0.28.1
redis==6.2.0
python-multipart==0.0.20
